backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-environ>=0.11.2
django-filter>=23.5

# Database
psycopg2-binary>=2.9.9
pymongo>=4.6.0
djongo>=1.3.6

# File Processing
Pillow>=10.2.0
python-magic>=0.4.27
PyPDF2>=3.0.1
pdf2image>=1.16.3

# API Clients
requests>=2.31.0
httpx>=0.26.0
openai>=1.10.0

# Data Processing
pandas>=2.1.4
openpyxl>=3.1.2
python-dateutil>=2.8.2

# Utilities
python-dotenv>=1.0.0
celery>=5.3.4
redis>=5.0.1

# Development
black>=23.12.1
flake8>=7.0.0
pytest>=7.4.4
pytest-django>=4.7.0
